third_party/blink/web_tests/fast/filesystem/op-get-metadata-expected.txt

* Running: GetMetadata
PASS Succeeded: "/".getMetadata()
PASS Succeeded: "/".getDirectory("/a")
PASS Succeeded: "/a".getMetadata()
PASS this.environment["/a.returned.modificationTime"] is >= this.roundedStartDate
PASS Succeeded: "/file1".getMetadata()
PASS this.environment["/file1.returned.modificationTime"] is >= this.roundedStartDate
PASS Succeeded: "/file2".getMetadata()
PASS this.environment["/file2.returned.modificationTime"] is >= this.roundedStartDate
PASS this.environment["/file2.size"] is this.environment["/file2.returned.size"]
PASS Succeeded: "/file3".getMetadata()
PASS this.environment["/file3.returned.modificationTime"] is >= this.roundedStartDate
PASS this.environment["/file3.size"] is this.environment["/file3.returned.size"]
PASS Succeeded: "/".getFile("/b")
PASS Succeeded: "/b".getMetadata()
PASS this.environment["/b.returned.modificationTime"] is >= this.roundedStartDate
PASS Succeeded: "/tmp".remove()
PASS this.expectedError + "" is "NotFoundError"
PASS this.environment["/a.returned.modificationTime"] is >= this.environment["/.returned.modificationTime"]
PASS this.environment["/b.returned.modificationTime"] is >= this.environment["/.returned.modificationTime"]
PASS this.environment["/b.returned.modificationTime"] is >= this.environment["/a.returned.modificationTime"]
Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
